pruebas de image.service, filterimages.pipe y app.component actualizadas
