+++ src/app/image.service.spec.ts
@@ -6,11 +6,29 @@ describe('ImageService', () => {
   let service: ImageService;
 
   beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(ImageService);
+    service = new ImageService();
   });
 
-  it('should be created', () => {
+  it('debe crearse el servicio', () => {
     expect(service).toBeTruthy();
+  });
+
+  describe('getImages', () => {
+    it('debe retornar todas las imagenes', () => {
+      let imagenes = service.getImages();
+      expect(imagenes.length).toEqual(5);
+    });
+  });
+
+  describe('getImage', () => {
+    it('debe retornar la imagen con id si existe', () => {
+      let imagene = service.getImage(1);
+      expect(imagene.brand).toEqual('perro');
+    });
+
+    it('debe retornar indefinido si se busca una imagen con id que NO existe', () => {
+      let imagene = service.getImage(100);
+      expect(imagene).toEqual(undefined);
+    });
   });
 });
